Update __init__.py
Updated the "HDR Lightmap" and "HDR Lightmap with Alpha" material templates. The albedo image texture + UV nodes now have a corresponding label applied to them. In addition, the HDR lightmap is now connected to the second color slot of the "Mix" node, instead of the first slot. The first slot is now connected to the albedo image texture instead.